+++ packages/jsts/src/analysis/projectAnalysis/analyzeProject.ts
