+++ packages/jsts/src/analysis/projectAnalysis/analyzeWithoutProgram.ts
@@ -22,7 +22,7 @@ import { relative } from 'node:path/posix';
 import { ProgressReport } from '../../../../shared/src/helpers/progress-report.js';
 import { handleFileResult } from './handleFileResult.js';
 import { WsIncrementalResult } from '../../../../bridge/src/request.js';
-import { isAnalysisCancelled } from './projectAnalyzer.js';
+import { isAnalysisCancelled } from './analyzeProject.js';
 
 /**
  * Analyzes files without type-checking.
